perf(server): fix KV cache dtype (bfloat16) + N-step graph batching + async_eval
- KV pool was allocating float16 but model weights are bfloat16 — dtype
  mismatch forced conversion kernels in every attention call (+56% C=1)
- Added N-step graph batching (N=16) to scheduler decode loop
- Added int-offset model.forward() overload to avoid mx::eval sync
- Switched to mx::async_eval for token evaluation

Results (Qwen3-0.6B):
  C=1: 73 → 114 tok/s (+56%)
  C=8: 165 → 190 tok/s (+15%, matches LM Studio)

+++ server/src/kv_pool.cpp
@@ -25,9 +25,9 @@ KVCachePool::KVCachePool(int max_slots, int max_context_len, int num_layers,
 
         for (int l = 0; l < num_layers; ++l) {
             cache_keys_[s].emplace_back(
-                mx::zeros({1, n_kv_heads, max_context_len, head_dim}, mx::float16));
+                mx::zeros({1, n_kv_heads, max_context_len, head_dim}, mx::bfloat16));
             cache_values_[s].emplace_back(
-                mx::zeros({1, n_kv_heads, max_context_len, head_dim}, mx::float16));
+                mx::zeros({1, n_kv_heads, max_context_len, head_dim}, mx::bfloat16));
 
             all_arrays.push_back(cache_keys_[s][l]);
             all_arrays.push_back(cache_values_[s][l]);
